+++ src/Gate.IO.Api/Futures/Enums/GateFuturesTriggerCondition.cs
@@ -6,6 +6,12 @@
 public enum GateFuturesTriggerCondition : byte
 {
     /// <summary>
+    /// No trigger condition is set.
+    /// </summary>
+    [Map("0")]
+    None = 0,
+
+    /// <summary>
     /// Greater than or equal to trigger condition
     /// </summary>
     [Map(">=")]